tidy(debezium): extract CDC payload helpers
Both DebeziumKafkaSource and MessageFormat.Json inline the same shape of
Debezium JSON envelope handling: unwrap `{schema, payload}` if present,
walk the tree converting JsonElements to plain JVM values, route the
resulting map into XTDB by `source.schema`/`source.table`.

A DebeziumEngineSource (embedded connector, no Kafka) reads the same
envelope shape and wants the same apply path.
Better to have one copy than three.

Pure refactor — no behaviour change.
`MessageFormat.Json.decode(ByteArray)` now delegates to the shared
`parseCdcEnvelope(String)`.
DebeziumKafkaSource's private writeCdcPayload / parseValidTimeMicros
become top-level in the same package.

Separated from the feat commit so the review can focus on what actually
moves.
